fix: update model to correctly reflect real-world

+++ src/MFEndpoint.Domain/GeoLocation.cs
@@ -2,6 +2,6 @@ namespace MFEndpoint.Domain;
 
 public class GeoLocation
 {
-    public double Latitude { get; set; }
-    public double Longitude { get; set; }
+    public Decimal Latitude { get; set; }
+    public Decimal Longitude { get; set; }
 }
